fix: 레시피당 재료 데이터베이스 추가 에러 수정

+++ app/src/main/java/com/example/recipe_provider/popups/AddRecipePopupActivity.java
@@ -32,7 +32,7 @@ import java.io.IOException;
 import java.util.HashMap;
 
 public class AddRecipePopupActivity extends Activity {
-    public int ingredientId;
+    public long ingredientId;
     public Integer amount_max;
     public Integer PICK_IMAGE_REQUEST = 1;
     public String imagePath = null;
